requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-jwt-extended>=4.5.0
flask-limiter>=3.5.0
flask-migrate>=4.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
marshmallow>=3.20.0
firebase-admin>=6.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
